refactor(rust): split `SecureChannel` into `Self` and `SecureChannelListener`
