+++ implementations/rust/ockam/ockam_command/src/secure_channel/create.rs
@@ -9,7 +9,7 @@ use ockam_api::error::ApiError;
 use ockam_api::nodes::models::secure_channel::CreateSecureChannelResponse;
 use ockam_api::nodes::NODEMANAGER_ADDR;
 use ockam_api::{route_to_multiaddr, Response, Status};
-use ockam_core::{route, Address, Route};
+use ockam_core::{route, Route};
 use ockam_multiaddr::MultiAddr;
 use tracing::debug;
 
@@ -20,18 +20,6 @@ pub struct CreateCommand {
 
     /// What address to connect to
     addr: MultiAddr,
-    /// Pre-known Identifiers of the other side
-    #[clap(short, long)]
-    authorized_identifier: Option<Vec<IdentityIdentifier>>,
-}
-
-#[derive(Clone, Debug, Args)]
-pub struct CreateListenerCommand {
-    #[clap(flatten)]
-    node_opts: NodeOpts,
-
-    /// Specify an address for this listener
-    bind: Address,
     /// Pre-known Identifiers of the other side
     #[clap(short, long)]
     authorized_identifier: Option<Vec<IdentityIdentifier>>,
@@ -49,23 +37,6 @@ impl CreateCommand {
         };
 
         connect_to(port, command, create_connector);
-
-        Ok(())
-    }
-}
-
-impl CreateListenerCommand {
-    pub fn run(opts: CommandGlobalOpts, command: CreateListenerCommand) -> anyhow::Result<()> {
-        let cfg = opts.config;
-        let port = match cfg.select_node(&command.node_opts.api_node) {
-            Some(cfg) => cfg.port,
-            None => {
-                eprintln!("No such node available.  Run `ockam node list` to list available nodes");
-                std::process::exit(-1);
-            }
-        };
-
-        connect_to(port, command, create_listener);
 
         Ok(())
     }
@@ -117,38 +88,6 @@ pub async fn create_connector(
         Ok(o) => println!("{o}"),
         Err(err) => eprintln!("{err}"),
     };
-
-    stop_node(ctx).await
-}
-
-pub async fn create_listener(
-    ctx: ockam::Context,
-    cmd: CreateListenerCommand,
-    mut base_route: Route,
-) -> anyhow::Result<()> {
-    let CreateListenerCommand {
-        bind: addr,
-        authorized_identifier: authorized_identifiers,
-        ..
-    } = cmd;
-
-    let resp: Vec<u8> = ctx
-        .send_and_receive(
-            base_route.modify().append(NODEMANAGER_ADDR),
-            api::create_secure_channel_listener(&addr, authorized_identifiers)?,
-        )
-        .await?;
-
-    let response = api::parse_create_secure_channel_listener_response(&resp)?;
-
-    match response.status() {
-        Some(Status::Ok) => {
-            eprintln!("Secure Channel Listener created at {}!", addr)
-        }
-        _ => {
-            eprintln!("An error occurred while creating secure channel listener",)
-        }
-    }
 
     stop_node(ctx).await
 }
